Migrate S06 relay-only test to `test_send_e2e.py` and refactor with `StepsCommon` for reusability.
